LPD-75170 refactor: rename ai-hub-rest-client to ai-hub-rest-nexus-client
