--- modules/dxp/apps/ai-hub-nexus/ai-hub-rest-client/bnd.bnd
@@ -1,3 +1,0 @@
-Bundle-Name: Liferay AI Hub REST Client
-Bundle-SymbolicName: com.liferay.ai.hub.rest.client
-Bundle-Version: 1.0.6
