--- modules/dxp/apps/ai-hub-nexus/ai-hub-rest-client/src/main/java/com/liferay/ai/hub/rest/client/aggregation/Aggregation.java
@@ -1,30 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2025 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.ai.hub.rest.client.aggregation;
-
-import jakarta.annotation.Generated;
-
-import java.util.HashMap;
-import java.util.Map;
-
-/**
- * @author Feliphe Marinho
- * @generated
- */
-@Generated("")
-public class Aggregation {
-
-	public Map<String, String> getAggregationTerms() {
-		return _aggregationTerms;
-	}
-
-	public void setAggregationTerms(Map<String, String> aggregationTerms) {
-		_aggregationTerms = aggregationTerms;
-	}
-
-	private Map<String, String> _aggregationTerms = new HashMap<>();
-
-}
