--- modules/dxp/apps/ai-hub-nexus/ai-hub-rest-client/src/main/java/com/liferay/ai/hub/rest/client/dto/v1_0/AgentDefinition.java
@@ -1,291 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2026 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.ai.hub.rest.client.dto.v1_0;
-
-import com.liferay.ai.hub.rest.client.function.UnsafeSupplier;
-import com.liferay.ai.hub.rest.client.serdes.v1_0.AgentDefinitionSerDes;
-
-import jakarta.annotation.Generated;
-
-import java.io.Serializable;
-
-import java.util.Map;
-import java.util.Objects;
-
-/**
- * @author Feliphe Marinho
- * @generated
- */
-@Generated("")
-public class AgentDefinition implements Cloneable, Serializable {
-
-	public static AgentDefinition toDTO(String json) {
-		return AgentDefinitionSerDes.toDTO(json);
-	}
-
-	public Map<String, Map<String, String>> getActions() {
-		return actions;
-	}
-
-	public void setActions(Map<String, Map<String, String>> actions) {
-		this.actions = actions;
-	}
-
-	public void setActions(
-		UnsafeSupplier<Map<String, Map<String, String>>, Exception>
-			actionsUnsafeSupplier) {
-
-		try {
-			actions = actionsUnsafeSupplier.get();
-		}
-		catch (Exception e) {
-			throw new RuntimeException(e);
-		}
-	}
-
-	protected Map<String, Map<String, String>> actions;
-
-	public Boolean getActive() {
-		return active;
-	}
-
-	public void setActive(Boolean active) {
-		this.active = active;
-	}
-
-	public void setActive(
-		UnsafeSupplier<Boolean, Exception> activeUnsafeSupplier) {
-
-		try {
-			active = activeUnsafeSupplier.get();
-		}
-		catch (Exception e) {
-			throw new RuntimeException(e);
-		}
-	}
-
-	protected Boolean active;
-
-	public String getDescription() {
-		return description;
-	}
-
-	public void setDescription(String description) {
-		this.description = description;
-	}
-
-	public void setDescription(
-		UnsafeSupplier<String, Exception> descriptionUnsafeSupplier) {
-
-		try {
-			description = descriptionUnsafeSupplier.get();
-		}
-		catch (Exception e) {
-			throw new RuntimeException(e);
-		}
-	}
-
-	protected String description;
-
-	public String getExternalReferenceCode() {
-		return externalReferenceCode;
-	}
-
-	public void setExternalReferenceCode(String externalReferenceCode) {
-		this.externalReferenceCode = externalReferenceCode;
-	}
-
-	public void setExternalReferenceCode(
-		UnsafeSupplier<String, Exception> externalReferenceCodeUnsafeSupplier) {
-
-		try {
-			externalReferenceCode = externalReferenceCodeUnsafeSupplier.get();
-		}
-		catch (Exception e) {
-			throw new RuntimeException(e);
-		}
-	}
-
-	protected String externalReferenceCode;
-
-	public Long getId() {
-		return id;
-	}
-
-	public void setId(Long id) {
-		this.id = id;
-	}
-
-	public void setId(UnsafeSupplier<Long, Exception> idUnsafeSupplier) {
-		try {
-			id = idUnsafeSupplier.get();
-		}
-		catch (Exception e) {
-			throw new RuntimeException(e);
-		}
-	}
-
-	protected Long id;
-
-	public Variable[] getInputVariables() {
-		return inputVariables;
-	}
-
-	public void setInputVariables(Variable[] inputVariables) {
-		this.inputVariables = inputVariables;
-	}
-
-	public void setInputVariables(
-		UnsafeSupplier<Variable[], Exception> inputVariablesUnsafeSupplier) {
-
-		try {
-			inputVariables = inputVariablesUnsafeSupplier.get();
-		}
-		catch (Exception e) {
-			throw new RuntimeException(e);
-		}
-	}
-
-	protected Variable[] inputVariables;
-
-	public Variable getOutputVariable() {
-		return outputVariable;
-	}
-
-	public void setOutputVariable(Variable outputVariable) {
-		this.outputVariable = outputVariable;
-	}
-
-	public void setOutputVariable(
-		UnsafeSupplier<Variable, Exception> outputVariableUnsafeSupplier) {
-
-		try {
-			outputVariable = outputVariableUnsafeSupplier.get();
-		}
-		catch (Exception e) {
-			throw new RuntimeException(e);
-		}
-	}
-
-	protected Variable outputVariable;
-
-	public Status getStatus() {
-		return status;
-	}
-
-	public void setStatus(Status status) {
-		this.status = status;
-	}
-
-	public void setStatus(
-		UnsafeSupplier<Status, Exception> statusUnsafeSupplier) {
-
-		try {
-			status = statusUnsafeSupplier.get();
-		}
-		catch (Exception e) {
-			throw new RuntimeException(e);
-		}
-	}
-
-	protected Status status;
-
-	public String getTitle() {
-		return title;
-	}
-
-	public void setTitle(String title) {
-		this.title = title;
-	}
-
-	public void setTitle(
-		UnsafeSupplier<String, Exception> titleUnsafeSupplier) {
-
-		try {
-			title = titleUnsafeSupplier.get();
-		}
-		catch (Exception e) {
-			throw new RuntimeException(e);
-		}
-	}
-
-	protected String title;
-
-	public Integer getVersion() {
-		return version;
-	}
-
-	public void setVersion(Integer version) {
-		this.version = version;
-	}
-
-	public void setVersion(
-		UnsafeSupplier<Integer, Exception> versionUnsafeSupplier) {
-
-		try {
-			version = versionUnsafeSupplier.get();
-		}
-		catch (Exception e) {
-			throw new RuntimeException(e);
-		}
-	}
-
-	protected Integer version;
-
-	public String getWorkflowDefinitionName() {
-		return workflowDefinitionName;
-	}
-
-	public void setWorkflowDefinitionName(String workflowDefinitionName) {
-		this.workflowDefinitionName = workflowDefinitionName;
-	}
-
-	public void setWorkflowDefinitionName(
-		UnsafeSupplier<String, Exception>
-			workflowDefinitionNameUnsafeSupplier) {
-
-		try {
-			workflowDefinitionName = workflowDefinitionNameUnsafeSupplier.get();
-		}
-		catch (Exception e) {
-			throw new RuntimeException(e);
-		}
-	}
-
-	protected String workflowDefinitionName;
-
-	@Override
-	public AgentDefinition clone() throws CloneNotSupportedException {
-		return (AgentDefinition)super.clone();
-	}
-
-	@Override
-	public boolean equals(Object object) {
-		if (this == object) {
-			return true;
-		}
-
-		if (!(object instanceof AgentDefinition)) {
-			return false;
-		}
-
-		AgentDefinition agentDefinition = (AgentDefinition)object;
-
-		return Objects.equals(toString(), agentDefinition.toString());
-	}
-
-	@Override
-	public int hashCode() {
-		String string = toString();
-
-		return string.hashCode();
-	}
-
-	public String toString() {
-		return AgentDefinitionSerDes.toJSON(this);
-	}
-
-}
