--- modules/dxp/apps/ai-hub-nexus/ai-hub-rest-client/src/main/java/com/liferay/ai/hub/rest/client/dto/v1_0/Scope.java
@@ -1,80 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2025 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.ai.hub.rest.client.dto.v1_0;
-
-import com.liferay.ai.hub.rest.client.function.UnsafeSupplier;
-import com.liferay.ai.hub.rest.client.serdes.v1_0.ScopeSerDes;
-
-import jakarta.annotation.Generated;
-
-import java.io.Serializable;
-
-import java.util.Objects;
-
-/**
- * @author Feliphe Marinho
- * @generated
- */
-@Generated("")
-public class Scope implements Cloneable, Serializable {
-
-	public static Scope toDTO(String json) {
-		return ScopeSerDes.toDTO(json);
-	}
-
-	public String getExternalReferenceCode() {
-		return externalReferenceCode;
-	}
-
-	public void setExternalReferenceCode(String externalReferenceCode) {
-		this.externalReferenceCode = externalReferenceCode;
-	}
-
-	public void setExternalReferenceCode(
-		UnsafeSupplier<String, Exception> externalReferenceCodeUnsafeSupplier) {
-
-		try {
-			externalReferenceCode = externalReferenceCodeUnsafeSupplier.get();
-		}
-		catch (Exception e) {
-			throw new RuntimeException(e);
-		}
-	}
-
-	protected String externalReferenceCode;
-
-	@Override
-	public Scope clone() throws CloneNotSupportedException {
-		return (Scope)super.clone();
-	}
-
-	@Override
-	public boolean equals(Object object) {
-		if (this == object) {
-			return true;
-		}
-
-		if (!(object instanceof Scope)) {
-			return false;
-		}
-
-		Scope scope = (Scope)object;
-
-		return Objects.equals(toString(), scope.toString());
-	}
-
-	@Override
-	public int hashCode() {
-		String string = toString();
-
-		return string.hashCode();
-	}
-
-	public String toString() {
-		return ScopeSerDes.toJSON(this);
-	}
-
-}
