--- modules/dxp/apps/ai-hub-nexus/ai-hub-rest-client/src/main/java/com/liferay/ai/hub/rest/client/dto/v1_0/Task.java
@@ -1,142 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2025 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.ai.hub.rest.client.dto.v1_0;
-
-import com.liferay.ai.hub.rest.client.function.UnsafeSupplier;
-import com.liferay.ai.hub.rest.client.serdes.v1_0.TaskSerDes;
-
-import jakarta.annotation.Generated;
-
-import java.io.Serializable;
-
-import java.util.Map;
-import java.util.Objects;
-
-/**
- * @author Feliphe Marinho
- * @generated
- */
-@Generated("")
-public class Task implements Cloneable, Serializable {
-
-	public static Task toDTO(String json) {
-		return TaskSerDes.toDTO(json);
-	}
-
-	public Map<String, ?> getContext() {
-		return context;
-	}
-
-	public void setContext(Map<String, ?> context) {
-		this.context = context;
-	}
-
-	public void setContext(
-		UnsafeSupplier<Map<String, ?>, Exception> contextUnsafeSupplier) {
-
-		try {
-			context = contextUnsafeSupplier.get();
-		}
-		catch (Exception e) {
-			throw new RuntimeException(e);
-		}
-	}
-
-	protected Map<String, ?> context;
-
-	public String getExternalReferenceCode() {
-		return externalReferenceCode;
-	}
-
-	public void setExternalReferenceCode(String externalReferenceCode) {
-		this.externalReferenceCode = externalReferenceCode;
-	}
-
-	public void setExternalReferenceCode(
-		UnsafeSupplier<String, Exception> externalReferenceCodeUnsafeSupplier) {
-
-		try {
-			externalReferenceCode = externalReferenceCodeUnsafeSupplier.get();
-		}
-		catch (Exception e) {
-			throw new RuntimeException(e);
-		}
-	}
-
-	protected String externalReferenceCode;
-
-	public String getSseEventSinkKey() {
-		return sseEventSinkKey;
-	}
-
-	public void setSseEventSinkKey(String sseEventSinkKey) {
-		this.sseEventSinkKey = sseEventSinkKey;
-	}
-
-	public void setSseEventSinkKey(
-		UnsafeSupplier<String, Exception> sseEventSinkKeyUnsafeSupplier) {
-
-		try {
-			sseEventSinkKey = sseEventSinkKeyUnsafeSupplier.get();
-		}
-		catch (Exception e) {
-			throw new RuntimeException(e);
-		}
-	}
-
-	protected String sseEventSinkKey;
-
-	public String getType() {
-		return type;
-	}
-
-	public void setType(String type) {
-		this.type = type;
-	}
-
-	public void setType(UnsafeSupplier<String, Exception> typeUnsafeSupplier) {
-		try {
-			type = typeUnsafeSupplier.get();
-		}
-		catch (Exception e) {
-			throw new RuntimeException(e);
-		}
-	}
-
-	protected String type;
-
-	@Override
-	public Task clone() throws CloneNotSupportedException {
-		return (Task)super.clone();
-	}
-
-	@Override
-	public boolean equals(Object object) {
-		if (this == object) {
-			return true;
-		}
-
-		if (!(object instanceof Task)) {
-			return false;
-		}
-
-		Task task = (Task)object;
-
-		return Objects.equals(toString(), task.toString());
-	}
-
-	@Override
-	public int hashCode() {
-		String string = toString();
-
-		return string.hashCode();
-	}
-
-	public String toString() {
-		return TaskSerDes.toJSON(this);
-	}
-
-}
