--- modules/dxp/apps/ai-hub-nexus/ai-hub-rest-client/src/main/java/com/liferay/ai/hub/rest/client/serdes/v1_0/ChatSerDes.java
@@ -1,202 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2025 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.ai.hub.rest.client.serdes.v1_0;
-
-import com.liferay.ai.hub.rest.client.dto.v1_0.Chat;
-import com.liferay.ai.hub.rest.client.json.BaseJSONParser;
-
-import jakarta.annotation.Generated;
-
-import java.util.Iterator;
-import java.util.Map;
-import java.util.Objects;
-import java.util.Set;
-import java.util.TreeMap;
-
-/**
- * @author Feliphe Marinho
- * @generated
- */
-@Generated("")
-public class ChatSerDes {
-
-	public static Chat toDTO(String json) {
-		ChatJSONParser chatJSONParser = new ChatJSONParser();
-
-		return chatJSONParser.parseToDTO(json);
-	}
-
-	public static Chat[] toDTOs(String json) {
-		ChatJSONParser chatJSONParser = new ChatJSONParser();
-
-		return chatJSONParser.parseToDTOs(json);
-	}
-
-	public static String toJSON(Chat chat) {
-		if (chat == null) {
-			return "null";
-		}
-
-		StringBuilder sb = new StringBuilder();
-
-		sb.append("{");
-
-		if (chat.getExternalReferenceCode() != null) {
-			if (sb.length() > 1) {
-				sb.append(", ");
-			}
-
-			sb.append("\"externalReferenceCode\": ");
-
-			sb.append("\"");
-
-			sb.append(_escape(chat.getExternalReferenceCode()));
-
-			sb.append("\"");
-		}
-
-		sb.append("}");
-
-		return sb.toString();
-	}
-
-	public static Map<String, Object> toMap(String json) {
-		ChatJSONParser chatJSONParser = new ChatJSONParser();
-
-		return chatJSONParser.parseToMap(json);
-	}
-
-	public static Map<String, String> toMap(Chat chat) {
-		if (chat == null) {
-			return null;
-		}
-
-		Map<String, String> map = new TreeMap<>();
-
-		if (chat.getExternalReferenceCode() == null) {
-			map.put("externalReferenceCode", null);
-		}
-		else {
-			map.put(
-				"externalReferenceCode",
-				String.valueOf(chat.getExternalReferenceCode()));
-		}
-
-		return map;
-	}
-
-	public static class ChatJSONParser extends BaseJSONParser<Chat> {
-
-		@Override
-		protected Chat createDTO() {
-			return new Chat();
-		}
-
-		@Override
-		protected Chat[] createDTOArray(int size) {
-			return new Chat[size];
-		}
-
-		@Override
-		protected boolean parseMaps(String jsonParserFieldName) {
-			if (Objects.equals(jsonParserFieldName, "externalReferenceCode")) {
-				return false;
-			}
-
-			return false;
-		}
-
-		@Override
-		protected void setField(
-			Chat chat, String jsonParserFieldName,
-			Object jsonParserFieldValue) {
-
-			if (Objects.equals(jsonParserFieldName, "externalReferenceCode")) {
-				if (jsonParserFieldValue != null) {
-					chat.setExternalReferenceCode((String)jsonParserFieldValue);
-				}
-			}
-		}
-
-	}
-
-	private static String _escape(Object object) {
-		String string = String.valueOf(object);
-
-		for (String[] strings : BaseJSONParser.JSON_ESCAPE_STRINGS) {
-			string = string.replace(strings[0], strings[1]);
-		}
-
-		return string;
-	}
-
-	private static String _toJSON(Map<String, ?> map) {
-		StringBuilder sb = new StringBuilder("{");
-
-		@SuppressWarnings("unchecked")
-		Set set = map.entrySet();
-
-		@SuppressWarnings("unchecked")
-		Iterator<Map.Entry<String, ?>> iterator = set.iterator();
-
-		while (iterator.hasNext()) {
-			Map.Entry<String, ?> entry = iterator.next();
-
-			sb.append("\"");
-			sb.append(entry.getKey());
-			sb.append("\": ");
-
-			Object value = entry.getValue();
-
-			sb.append(_toJSON(value));
-
-			if (iterator.hasNext()) {
-				sb.append(", ");
-			}
-		}
-
-		sb.append("}");
-
-		return sb.toString();
-	}
-
-	private static String _toJSON(Object value) {
-		if (value == null) {
-			return "null";
-		}
-
-		if (value instanceof Map) {
-			return _toJSON((Map)value);
-		}
-
-		Class<?> clazz = value.getClass();
-
-		if (clazz.isArray()) {
-			StringBuilder sb = new StringBuilder("[");
-
-			Object[] values = (Object[])value;
-
-			for (int i = 0; i < values.length; i++) {
-				sb.append(_toJSON(values[i]));
-
-				if ((i + 1) < values.length) {
-					sb.append(", ");
-				}
-			}
-
-			sb.append("]");
-
-			return sb.toString();
-		}
-
-		if (value instanceof String) {
-			return "\"" + _escape(value) + "\"";
-		}
-
-		return String.valueOf(value);
-	}
-
-}
