--- modules/dxp/apps/ai-hub-nexus/ai-hub-rest-client/src/main/java/com/liferay/ai/hub/rest/client/serdes/v1_0/VariableSerDes.java
@@ -1,229 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2026 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.ai.hub.rest.client.serdes.v1_0;
-
-import com.liferay.ai.hub.rest.client.dto.v1_0.Variable;
-import com.liferay.ai.hub.rest.client.json.BaseJSONParser;
-
-import jakarta.annotation.Generated;
-
-import java.util.Iterator;
-import java.util.Map;
-import java.util.Objects;
-import java.util.Set;
-import java.util.TreeMap;
-
-/**
- * @author Feliphe Marinho
- * @generated
- */
-@Generated("")
-public class VariableSerDes {
-
-	public static Variable toDTO(String json) {
-		VariableJSONParser variableJSONParser = new VariableJSONParser();
-
-		return variableJSONParser.parseToDTO(json);
-	}
-
-	public static Variable[] toDTOs(String json) {
-		VariableJSONParser variableJSONParser = new VariableJSONParser();
-
-		return variableJSONParser.parseToDTOs(json);
-	}
-
-	public static String toJSON(Variable variable) {
-		if (variable == null) {
-			return "null";
-		}
-
-		StringBuilder sb = new StringBuilder();
-
-		sb.append("{");
-
-		if (variable.getName() != null) {
-			if (sb.length() > 1) {
-				sb.append(", ");
-			}
-
-			sb.append("\"name\": ");
-
-			sb.append("\"");
-
-			sb.append(_escape(variable.getName()));
-
-			sb.append("\"");
-		}
-
-		if (variable.getType() != null) {
-			if (sb.length() > 1) {
-				sb.append(", ");
-			}
-
-			sb.append("\"type\": ");
-
-			sb.append("\"");
-
-			sb.append(_escape(variable.getType()));
-
-			sb.append("\"");
-		}
-
-		sb.append("}");
-
-		return sb.toString();
-	}
-
-	public static Map<String, Object> toMap(String json) {
-		VariableJSONParser variableJSONParser = new VariableJSONParser();
-
-		return variableJSONParser.parseToMap(json);
-	}
-
-	public static Map<String, String> toMap(Variable variable) {
-		if (variable == null) {
-			return null;
-		}
-
-		Map<String, String> map = new TreeMap<>();
-
-		if (variable.getName() == null) {
-			map.put("name", null);
-		}
-		else {
-			map.put("name", String.valueOf(variable.getName()));
-		}
-
-		if (variable.getType() == null) {
-			map.put("type", null);
-		}
-		else {
-			map.put("type", String.valueOf(variable.getType()));
-		}
-
-		return map;
-	}
-
-	public static class VariableJSONParser extends BaseJSONParser<Variable> {
-
-		@Override
-		protected Variable createDTO() {
-			return new Variable();
-		}
-
-		@Override
-		protected Variable[] createDTOArray(int size) {
-			return new Variable[size];
-		}
-
-		@Override
-		protected boolean parseMaps(String jsonParserFieldName) {
-			if (Objects.equals(jsonParserFieldName, "name")) {
-				return false;
-			}
-			else if (Objects.equals(jsonParserFieldName, "type")) {
-				return false;
-			}
-
-			return false;
-		}
-
-		@Override
-		protected void setField(
-			Variable variable, String jsonParserFieldName,
-			Object jsonParserFieldValue) {
-
-			if (Objects.equals(jsonParserFieldName, "name")) {
-				if (jsonParserFieldValue != null) {
-					variable.setName((String)jsonParserFieldValue);
-				}
-			}
-			else if (Objects.equals(jsonParserFieldName, "type")) {
-				if (jsonParserFieldValue != null) {
-					variable.setType((String)jsonParserFieldValue);
-				}
-			}
-		}
-
-	}
-
-	private static String _escape(Object object) {
-		String string = String.valueOf(object);
-
-		for (String[] strings : BaseJSONParser.JSON_ESCAPE_STRINGS) {
-			string = string.replace(strings[0], strings[1]);
-		}
-
-		return string;
-	}
-
-	private static String _toJSON(Map<String, ?> map) {
-		StringBuilder sb = new StringBuilder("{");
-
-		@SuppressWarnings("unchecked")
-		Set set = map.entrySet();
-
-		@SuppressWarnings("unchecked")
-		Iterator<Map.Entry<String, ?>> iterator = set.iterator();
-
-		while (iterator.hasNext()) {
-			Map.Entry<String, ?> entry = iterator.next();
-
-			sb.append("\"");
-			sb.append(entry.getKey());
-			sb.append("\": ");
-
-			Object value = entry.getValue();
-
-			sb.append(_toJSON(value));
-
-			if (iterator.hasNext()) {
-				sb.append(", ");
-			}
-		}
-
-		sb.append("}");
-
-		return sb.toString();
-	}
-
-	private static String _toJSON(Object value) {
-		if (value == null) {
-			return "null";
-		}
-
-		if (value instanceof Map) {
-			return _toJSON((Map)value);
-		}
-
-		Class<?> clazz = value.getClass();
-
-		if (clazz.isArray()) {
-			StringBuilder sb = new StringBuilder("[");
-
-			Object[] values = (Object[])value;
-
-			for (int i = 0; i < values.length; i++) {
-				sb.append(_toJSON(values[i]));
-
-				if ((i + 1) < values.length) {
-					sb.append(", ");
-				}
-			}
-
-			sb.append("]");
-
-			return sb.toString();
-		}
-
-		if (value instanceof String) {
-			return "\"" + _escape(value) + "\"";
-		}
-
-		return String.valueOf(value);
-	}
-
-}
