+++ modules/dxp/apps/ai-hub-nexus/ai-hub-rest-nexus-client/bnd.bnd
@@ -1,0 +1,3 @@
+Bundle-Name: Liferay AI Hub Nexus REST Client
+Bundle-SymbolicName: com.liferay.ai.hub.nexus.rest.client
+Bundle-Version: 1.0.6
