+++ modules/dxp/apps/ai-hub-nexus/ai-hub-rest-nexus-client/src/main/java/com/liferay/ai/hub/nexus/rest/client/dto/v1_0/Message.java
@@ -1,0 +1,138 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2025 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+package com.liferay.ai.hub.nexus.rest.client.dto.v1_0;
+
+import com.liferay.ai.hub.nexus.rest.client.function.UnsafeSupplier;
+import com.liferay.ai.hub.nexus.rest.client.serdes.v1_0.MessageSerDes;
+
+import jakarta.annotation.Generated;
+
+import java.io.Serializable;
+
+import java.util.Map;
+import java.util.Objects;
+
+/**
+ * @author Feliphe Marinho
+ * @generated
+ */
+@Generated("")
+public class Message implements Cloneable, Serializable {
+
+	public static Message toDTO(String json) {
+		return MessageSerDes.toDTO(json);
+	}
+
+	public Chat getChat() {
+		return chat;
+	}
+
+	public void setChat(Chat chat) {
+		this.chat = chat;
+	}
+
+	public void setChat(UnsafeSupplier<Chat, Exception> chatUnsafeSupplier) {
+		try {
+			chat = chatUnsafeSupplier.get();
+		}
+		catch (Exception e) {
+			throw new RuntimeException(e);
+		}
+	}
+
+	protected Chat chat;
+
+	public Map<String, ?> getContext() {
+		return context;
+	}
+
+	public void setContext(Map<String, ?> context) {
+		this.context = context;
+	}
+
+	public void setContext(
+		UnsafeSupplier<Map<String, ?>, Exception> contextUnsafeSupplier) {
+
+		try {
+			context = contextUnsafeSupplier.get();
+		}
+		catch (Exception e) {
+			throw new RuntimeException(e);
+		}
+	}
+
+	protected Map<String, ?> context;
+
+	public Scope getScope() {
+		return scope;
+	}
+
+	public void setScope(Scope scope) {
+		this.scope = scope;
+	}
+
+	public void setScope(UnsafeSupplier<Scope, Exception> scopeUnsafeSupplier) {
+		try {
+			scope = scopeUnsafeSupplier.get();
+		}
+		catch (Exception e) {
+			throw new RuntimeException(e);
+		}
+	}
+
+	protected Scope scope;
+
+	public String getText() {
+		return text;
+	}
+
+	public void setText(String text) {
+		this.text = text;
+	}
+
+	public void setText(UnsafeSupplier<String, Exception> textUnsafeSupplier) {
+		try {
+			text = textUnsafeSupplier.get();
+		}
+		catch (Exception e) {
+			throw new RuntimeException(e);
+		}
+	}
+
+	protected String text;
+
+	@Override
+	public Message clone() throws CloneNotSupportedException {
+		return (Message)super.clone();
+	}
+
+	@Override
+	public boolean equals(Object object) {
+		if (this == object) {
+			return true;
+		}
+
+		if (!(object instanceof Message)) {
+			return false;
+		}
+
+		Message message = (Message)object;
+
+		return Objects.equals(toString(), message.toString());
+	}
+
+	@Override
+	public int hashCode() {
+		String string = toString();
+
+		return string.hashCode();
+	}
+
+	public String toString() {
+		return MessageSerDes.toJSON(this);
+	}
+
+}
